+++ src/Resources/contao/config/config.php
@@ -1,0 +1,4 @@
+<?php
+
+$GLOBALS['TL_HOOKS']['formhybridOnCreateInstance'][] = [
+    \HeimrichHannot\ContaoFormhybridCompatibilityBundle\EventListener\FormhybridOnCreateInstanceListener::class, '__invoke'];